1. Membuat forecast ringkas untuk hari ini dan 2 hari kedepan dengan menggunakan recycle view. 2. mengubah tipe data epochdate pada model ForecastDay. 3. menambahkan method get forecast weather pada interface weather data. 4. mengubah layout kasar. 5. menguah maiun activity untuk dapat emnggunakan recyle view.

+++ app/src/main/java/com/fwmubarok/myforecastweather/Model/ForecastDay.java
@@ -12,7 +12,7 @@ public class ForecastDay {
 
     @SerializedName("date_epoch")
     @Expose
-    private Integer dateEpoch;
+    private long dateEpoch;
 
     @SerializedName("day")
     @Expose
@@ -34,7 +34,7 @@ public class ForecastDay {
         this.date = date;
     }
 
-    public Integer getDateEpoch() {
+    public long getDateEpoch() {
         return dateEpoch;
     }
